+++ lib/core/include/core/network/Peer.hpp
@@ -26,7 +26,7 @@ public:
   void stop();
   int connect(const std::string&, int);
   int disconnect(int);
-  ssize_t send(int, const MessageType&, const char*, size_t);
+  int send(int, const MessageType&, const char*, size_t);
 };
 
 #endif
